projects/policyengine-api-full/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "policyengine-api-full"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    {name = "PolicyEngine", email = "hello@policyengine.org"},
]
license = {file = "../../LICENSE"}
requires-python = ">=3.13"
dependencies = [
    "opentelemetry-instrumentation-sqlalchemy (>=0.51b0,<0.52)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.51b0,<0.52)",
    "policyengine-fastapi",
    "policyengine[uk,us]>=3.0.0",
    "psycopg2-binary>=2.9.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/policyengine_api_full"]

[tool.uv.sources]
policyengine-fastapi = { path = "../../libs/policyengine-fastapi", editable = true }

[project.optional-dependencies]
test = [ "pytest>=8.3.4", "pytest-asyncio>=0.25.3", "pytest-cov>=6.1.1",]
build = [ "pyright>=1.1.401", "black>=25.1.0", "openapi-python-client>=0.21.6",]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = ["tests"]
addopts = "--cov=policyengine_api_full --cov-report=term-missing"
